More testing fixes
