+++ linchpin/tests/mockdata/dummy/conf/linchpin.conf
@@ -5,11 +5,13 @@
 
 [DEFAULT]
 # name of the python package (Redundant, but easier than programmatically
-# obtaining the value. It's very unlike to change.)
+# obtaining the value. It's very unlikely to change.)
 pkg = linchpin
 
 # travis-ci doesn't like ~/.config/linchpin
 default_config_path = /tmp
+
+external_providers_path = %(default_config_path)s/linchpin-x
 
 [lp]
 # load custom ansible modules from here
@@ -30,6 +32,8 @@ rundb_schema = {"action": "",
                 "uhash": ""}
 rundb_hash = sha256
 
+#dateformat = %%m/%%d/%%Y %%I:%%M:%%S %%p
+
 
 [evars]
 # extra vars to pass to the ansible playbook being executed
@@ -37,6 +41,7 @@ _check_mode = False
 _async = False
 async_timeout = 1000
 output = True
+default_ssh_key_path = ~/.ssh
 
 # default paths in playbooks
 #
@@ -48,11 +53,11 @@ output = True
 layouts_folder = layouts
 topologies_folder = topologies
 
-schemas_folder = schemas
-
-resources_folder = resources
-inventories_folder = inventories
 hooks_folder = hooks
+roles_folder = roles
+inventories_folder = inventories
+resources_folder = resources
+schemas_folder = schemas
 
 # used in the API/CLI only
 playbooks_folder = provision
@@ -61,16 +66,15 @@ playbooks_folder = provision
 default_schemas_path = {{ lp_path }}/defaults/%(schemas_folder)s
 default_topologies_path = {{ lp_path }}/defaults/%(topologies_folder)s
 default_layouts_path = {{ lp_path }}/defaults/%(layouts_folder)s
-
-# outputs
-default_resources_path = {{ lp_path }}/defaults/%(resources_folder)s # formerly outputs
 default_inventories_path = {{ lp_path }}/defaults/%(inventories_folder)s
+default_roles_path = {{ lp_path }}/%(playbooks_folder)s/%(roles_folder)s
 
 # default_schema interpolates the default_schemas_path from above
 schema_v3 = %(default_schemas_path)s/schema_v3.json
 schema_v4 = %(default_schemas_path)s/schema_v4.json
 
-#keystore_path: "{{ playbook_dir }}/../keystore"
+# default credentials data, this path doesn't automatically exist
+default_credentials_path = %(default_config_path)s
 
 [init]
 # source path of files generated by linchpin init
@@ -78,40 +82,64 @@ source = templates/
 # formal name of the PinFile. Can be changed as desired.
 pinfile = PinFile
 
-[playbooks]
-# possible playbook actions and the corresponding action file
-up = site.yml
-destroy = site.yml
-down = site.yml
-schema_check = schemacheck.yml
-inv_gen = invgen.yml
-test = test.yml
+#[playbooks]
+## possible playbook actions and the corresponding action file
+#up = site.yml
+#destroy = site.yml
+#down = site.yml
+#schema_check = schemacheck.yml
+#inv_gen = invgen.yml
+#test = test.yml
 
 [logger]
 enable = True
+file = linchpin.log
+#format = %%(levelname)s %%(asctime)s %%(message)s
+#dateformat = %%m/%%d/%%Y %%I:%%M:%%S %%p
 level = logging.DEBUG
 
 [console]
 # logging to the console should also be possible
 enable = True # DO NOT REMOVE. Placeholder only, cannot disable.
+#format = %%(message)s
 level = logging.INFO
 
 [hookstates]
-up = pre,post
-inv = pre
+up = pre,post,inv
+inv = post
 destroy = pre,post
 
 [extensions]
-resource = .output
+playbooks = .yml
 inventory = .inventory
 
 [ansible]
 console = False
 
-[core_workspace_dirs]
-credentials = True
-hooks = True
-inventories = True
+[states]
+# in future each state will carry a comma seperated substates
+preup = preup
+predestroy = predestroy
+postup = postup
+postdestroy = postdestroy
+postinv = inventory
+
+[repository_control]
+driver = github
+
+[fetch_types]
 layouts = True
-resources = True
 topologies = True
+hooks = True
+credentials = True
+PinFile = True
+resources = False
+inventory = False
+
+[fetch_aliases]
+layout = layouts
+topology = topologies
+hooks = hooks
+credentials = credentials
+pinfile = PinFile
+workspace = workspace
